% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_BCR.R
\name{bcr_pal}
\alias{bcr_pal}
\title{BCR color palette}
\usage{
bcr_pal()
}
\description{
The standard color palette for line plots
}
\seealso{
Other colour BCR: 
\code{\link{scale_colour_BCR}()}
}
\concept{colour BCR}
