% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_BCR.R
\docType{data}
\name{colors_BCR}
\alias{colors_BCR}
\title{BCR color palette}
\format{
An object of class \code{list} of length 11.
}
\usage{
colors_BCR
}
\description{
The standard color palette for line plots
}
\seealso{
Other colour BCR: 
\code{\link{scale_colour_BCR}()}
}
\concept{colour BCR}
\keyword{datasets}
