% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_BCR.R
\name{scale_colour_BCR}
\alias{scale_colour_BCR}
\alias{scale_color_BCR}
\alias{scale_fill_BCR}
\title{BCR color scales}
\usage{
scale_colour_BCR(palette = "default", n_colors = 1)

scale_color_BCR(palette = "default", n_colors = 1)

scale_fill_BCR(palette = "default", n_colors = 1)
}
\description{
Color scales using the colors in the BCR color palette.
}
\concept{colour BCR}
