% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_BCR.R
\name{themeBCR}
\alias{themeBCR}
\title{Custom ggplot2 Theme: BCR
Theme inspired by Bird Conservancy of the Rockies}
\usage{
themeBCR()
}
\description{
\href{https://birdconservancy.org}{Bird Conservancy of the Rockies}.
}
\concept{themes BCR}
