% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_BCR.R
\name{wingspan_pal}
\alias{wingspan_pal}
\title{Wingspan color palette}
\usage{
wingspan_pal()
}
\description{
The standard color palette for line plots based on the game Wingspan
}
\seealso{
Other colour Wingspan: 
\code{\link{scale_colour_wingspan}()}
}
\concept{colour Wingspan}
